+++ core/database/src/main/java/com/kazemieh/database/entity/FinancialSourceEntity.kt
@@ -9,5 +9,5 @@ import kotlinx.serialization.Serializable
 data class FinancialSourceEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val balance: Double = 0.0
+    val balance: Int = 0
 )
