+++ core/domain/src/main/java/com/kazemieh/domain/usecase/AddFinancialSource.kt
@@ -5,7 +5,7 @@ import com.kazemieh.domain.repository.TransactionRepository
 class AddFinancialSource(
     private val repository: TransactionRepository
 ) {
-    suspend operator fun invoke(financialSourceName: String, balance: Double = 0.0): Long {
+    suspend operator fun invoke(financialSourceName: String, balance: Int = 0): Long {
         return repository.insertFinancialSource(
             financialSourceName = financialSourceName,
             balance = balance
