+++ core/model/src/main/java/com/kazemieh/model/Transaction.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Transaction(
     val id: Long,
-    val amount: Double,
+    val amount: Int,
     val categoryId: Long,
     val financialSourceId: Long,
     val description: String? = null,
