+++ feature/FinancialSource/src/main/java/com/kazemieh/financialsource/ui/add/AddFinancialSourceViewModel.kt
@@ -2,7 +2,6 @@ package com.kazemieh.financialsource.ui.add
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kazemieh.domain.usecase.AddCategory
 import com.kazemieh.domain.usecase.AddFinancialSource
 import kotlinx.coroutines.launch
 
@@ -10,10 +9,10 @@ class AddFinancialSourceViewModel(
     private val addFinancialSourceUseCase: AddFinancialSource
 ) : ViewModel() {
 
-    fun addFinancialSource(name: String,balance: Double = 0.0) {
+    fun addFinancialSource(name: String, balance: Int = 0) {
         viewModelScope.launch {
             if (name.isNotBlank()) {
-                addFinancialSourceUseCase(name,balance)
+                addFinancialSourceUseCase(name, balance)
             }
         }
     }
